backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.8
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.6
alembic==1.13.1
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
pyotp==2.9.0
pydantic[email]==2.5.2 